Split init_fakes into files named for the production TU each fakes
Follow-up to the review of #10807 (comments on init_fakes.cc:1 and :122):
"the stubs in this file should be moved to a file named for the file where
the functions being stubbed live". init_fakes.cc had become the host for
every seam the init suite happened to need, including 42 control globals
that other fakes files consumed through hand-written extern re-declarations.

Convention for the files added here: snake_case <production_tu>_fakes.{cc,h},
matching hip_fakes / env_fakes / tuning_fakes. Existing *_stubs.cc files keep
their names; symbols move into them rather than a new file whenever the owning
TU already has one.

New files, one per owning TU:
  amdsmi_fakes        src/misc/amdsmi_wrap.cc
  api_trace_fakes     src/misc/api_trace.cc NCCL_API dispatch
  env_plugin_fakes    src/plugin/env.cc
  gin_fakes           src/plugin/gin.cc + src/gin/gin_host.cc
  group_fakes         src/group.cc
  init_nvtx_fakes     src/init_nvtx.cc
  kernel_config_fakes src/misc/kernel_config.cc
  libc_interposers    gethostname / dladdr (no RCCL TU owns these)
  mem_manager_fakes   src/mem_manager.cc
  rocmwrap_fakes      src/misc/rocmwrap.cc

Moved into the file that already held the stub for the symbol, turning each
extern re-declaration into a definition and giving it a header:
  bootstrap_stubs     the bootstrapAllGather / getUniqueId / bcastGrowHandle
                      seams and bootstrapNetInit
  topo_stubs          the src/graph/*.cc ladder (new topo_stubs.h), plus
                      ncclGetUserP2pLevel and ncclTreeBasePostset from
                      transport_stubs, which are paths.cc and connect.cc
  transport_stubs     the NVLS seams, ncclGpuGdrSupport, ncclParamNvlsEnable
  os_fakes            every src/os/linux.cc entry point and seam, including the
                      four that were in nccl_stubs.cc (new os_fakes.h); the file
                      joins both init source lists
  strongstream_stubs  ncclStrongStreamConstruct/Synchronize and
                      ncclCudaContextTrack (new strongstream_stubs.h)
  rccl_wrap_fakes     validHsaScratchEnvSetting, getFirmwareVersion,
                      rcclSetDefaultBuffSizes, rcclSetP2pNetChunkSize
  nccl_stubs          the seams whose stubs it already owned
  env_fakes           initEnv
  tuning_fakes        ncclParamPatEnable

Every relocated global's reset moves with it into that module's Reset*, which
ResetInitFakes now chains; init_fakes.h becomes the aggregation header, so
init-test.cc is unchanged.

Pure refactor: src/ is byte-identical, and both copies of each doubled CMake
target list were updated. Test counts unchanged (688 / 687 / 110 / 178 / 176 /
317), src/init.cc coverage unchanged at 89.74% lines and 69/69 functions,
--gtest_shuffle green on 5 seeds for both init arms.
