+++ projects/rccl/test/host/fakes/strongstream_stubs.cc
@@ -16,6 +16,9 @@
 #include "strongstream.h"
 
 #include "fail_loud.h"
+#include "strongstream_stubs.h"
+
+struct ncclCudaContext;
 
 ncclResult_t ncclCudaGetCapturingGraph(struct ncclCudaGraph*, hipStream_t, int) {
   FailLoudUnfaked("strongstream_stubs", "ncclCudaGetCapturingGraph");
@@ -32,3 +35,22 @@ ncclResult_t ncclStrongStreamAcquiredWorkStream(struct ncclCudaGraph, struct ncc
 }
 // Benign teardown: commFree reaches this on a happy-path destroy.
 ncclResult_t ncclStrongStreamDestruct(struct ncclStrongStream*) { return ncclSuccess; }
+
+ncclResult_t g_ncclStrongStreamResult = ncclSuccess;
+ncclResult_t ncclStrongStreamConstruct(struct ncclStrongStream*) { return g_ncclStrongStreamResult; }
+ncclResult_t ncclStrongStreamSynchronize(struct ncclStrongStream*) { return g_ncclStrongStreamResult; }
+
+// Controllable (was fail-loud in nccl_stubs.cc).
+ncclResult_t g_ncclCudaContextTrackResult = ncclSuccess;
+int g_ncclCudaContextTrackCalls = 0;
+ncclResult_t ncclCudaContextTrack(struct ncclCudaContext** out) {
+  g_ncclCudaContextTrackCalls++;
+  if (out) *out = nullptr;
+  return g_ncclCudaContextTrackResult;
+}
+
+void ResetStrongStreamStubs() {
+  g_ncclStrongStreamResult = ncclSuccess;
+  g_ncclCudaContextTrackResult = ncclSuccess;
+  g_ncclCudaContextTrackCalls = 0;
+}
