+++ projects/rccl/test/host/fakes/transport_stubs.cc
@@ -12,13 +12,28 @@
 
 #include "transport_stubs.h"
 
+#include <cstdint>
 #include <cstdlib>
 #include <functional>
 
 #include "nccl.h"
+#include "nccl_fakes.h"  // g_loadParam, for the NCCL_PARAM default this floor stands in for
 
 struct ncclComm;
 struct ncclTopoGraph;
+
+// src/transport/nvls.cc:159. Referenced by init.cc but not declared inside it, so the redirected
+// NCCL_PARAM does not cover it.
+int64_t ncclParamNvlsEnable() { return g_loadParam("NVLS_ENABLE", 2); }
+
+// src/plugin/net.cc: initTransportsRank's GDR probe.
+int g_gdrSupportValue = 0;
+int g_gdrSupportCalls = 0;
+ncclResult_t ncclGpuGdrSupport(struct ncclComm*, int* gdrSupport) {
+  ++g_gdrSupportCalls;
+  if (gdrSupport) *gdrSupport = g_gdrSupportValue;
+  return ncclSuccess;
+}
 
 // src/transport/net.cc:343. Was a fail-loud stub in nccl_stubs.cc, which forced
 // the enqueue target to omit it via a macro and supply its own; a seam here
@@ -34,21 +49,23 @@ std::function<ncclResult_t(struct ncclComm*)> g_ncclProxyStop = DefaultNcclProxy
 void ResetTransportStubs() {
   g_rcclUseAinic = false;
   g_ncclProxyStop = DefaultNcclProxyStop;
+  g_gdrSupportValue = 0;
+  g_gdrSupportCalls = 0;
+  g_ncclNvlsInitResult = ncclSuccess;
+  g_ncclNvlsInitCalls = 0;
+  g_ncclNvlsTuningResult = ncclSuccess;
+  g_ncclNvlsTuningCalls = 0;
 }
 
 ncclResult_t ncclCollNetChainBufferSetup(ncclComm_t comm) { ::abort(); }
 ncclResult_t ncclCollNetDirectBufferSetup(ncclComm_t comm) { ::abort(); }
 ncclResult_t ncclCollNetSetup(ncclComm_t comm, ncclComm_t parent, struct ncclTopoGraph* graphs[]) { ::abort(); }
-// Controllable (was fail-loud). A std::function, not a result code: initTransportsRank:1506 branches on the WRITTEN
-// *level, so a result-only seam could not drive it. Single call site (:1501) -> a success default is safe;
-// see MICROTEST_README.md "Adding more controllable seams" for when to default a seam to failure instead.
-extern std::function<ncclResult_t(int*)> g_ncclGetUserP2pLevel;
-ncclResult_t ncclGetUserP2pLevel(int* level) { return g_ncclGetUserP2pLevel(level); }
+// ncclGetUserP2pLevel (src/graph/paths.cc): topo_stubs.cc, with the rest of that TU's seams.
 ncclResult_t ncclNvlsBufferSetup(struct ncclComm* comm) { ::abort(); }
 // Controllable (was fail-loud). :1618 uses bare NCCLCHECK, not NCCLCHECKGOTO, so a failure here returns
 // WITHOUT running exit: -- the counter on ncclOsCpuCount is what makes that bypass observable.
-extern ncclResult_t g_ncclNvlsInitResult;
-extern int g_ncclNvlsInitCalls;
+ncclResult_t g_ncclNvlsInitResult = ncclSuccess;
+int g_ncclNvlsInitCalls = 0;
 ncclResult_t ncclNvlsInit(struct ncclComm* comm) {
   g_ncclNvlsInitCalls++;
   return g_ncclNvlsInitResult;
@@ -56,8 +73,8 @@ ncclResult_t ncclNvlsInit(struct ncclComm* comm) {
 ncclResult_t ncclNvlsSetup(struct ncclComm* comm, struct ncclComm* parent) { ::abort(); }
 ncclResult_t ncclNvlsTreeConnect(struct ncclComm* comm) { ::abort(); }
 // Controllable (was fail-loud). init.cc:2185, gated on comm->nvlsSupport surviving the :2182 fold.
-extern ncclResult_t g_ncclNvlsTuningResult;
-extern int g_ncclNvlsTuningCalls;
+ncclResult_t g_ncclNvlsTuningResult = ncclSuccess;
+int g_ncclNvlsTuningCalls = 0;
 ncclResult_t ncclNvlsTuning(struct ncclComm* comm) {
   g_ncclNvlsTuningCalls++;
   return g_ncclNvlsTuningResult;
@@ -70,15 +87,7 @@ int ncclPxnDisable(struct ncclComm* comm) { ::abort(); }
 ncclResult_t ncclTransportPatConnect(struct ncclComm* comm) { ::abort(); }
 ncclResult_t ncclTransportRingConnect(struct ncclComm* comm) { ::abort(); }
 ncclResult_t ncclTransportTreeConnect(struct ncclComm* comm) { ::abort(); }
-// Controllable (was fail-loud). init.cc:2215, gated on comm->topo->treeDefined.
-extern ncclResult_t g_ncclTreeBasePostsetResult;
-extern int g_ncclTreeBasePostsetCalls;
-extern struct ncclTopoGraph* g_ncclTreeBasePostsetGraph;
-ncclResult_t ncclTreeBasePostset(struct ncclComm* comm, struct ncclTopoGraph* treeGraph) {
-  g_ncclTreeBasePostsetCalls++;
-  g_ncclTreeBasePostsetGraph = treeGraph;
-  return g_ncclTreeBasePostsetResult;
-}
+// ncclTreeBasePostset (src/graph/connect.cc): topo_stubs.cc, next to ncclTopoPreset/ncclTopoPostset.
 ncclResult_t ncclTransportCheckP2pType(struct ncclComm*, bool*, bool*, bool*) { ::abort(); }
 ncclResult_t ncclTransportP2pConnect(struct ncclComm*, int, int, int*, int, int*, int) { ::abort(); }
 ncclResult_t ncclTransportP2pSetup(struct ncclComm*, struct ncclTopoGraph*, int, bool*) { ::abort(); }
