+++ projects/rccl/test/host/fakes/transport_stubs.h
@@ -8,12 +8,6 @@
 // re-declared by each consumer, so a signature change is a compile error at
 // every use rather than a silent link-time mismatch.
 //
-// SCOPE: this rule covers what transport_stubs.cc OWNS. It does not retroactively
-// cover the seams it merely CONSUMES -- g_ncclGetUserP2pLevel,
-// g_ncclNvlsInitResult and g_ncclNvlsInitCalls are defined in init_fakes.cc and
-// still hand-declared extern below, as they are in topo_stubs.cc, nccl_stubs.cc
-// and bootstrap_stubs.cc. Giving those a header means moving them to a fakes
-// file named after their owning production TU, which is a separate change.
 
 #ifndef RCCL_TEST_HOST_TRANSPORT_STUBS_H_
 #define RCCL_TEST_HOST_TRANSPORT_STUBS_H_
@@ -31,6 +25,18 @@ extern bool g_rcclUseAinic;
 
 // ncclProxyStop (src/proxy.cc): comm teardown stops the proxy through this.
 extern std::function<ncclResult_t(struct ncclComm*)> g_ncclProxyStop;
+
+// ncclGpuGdrSupport (src/plugin/net.cc): initTransportsRank's GDR probe.
+extern int g_gdrSupportValue;
+extern int g_gdrSupportCalls;
+
+// src/transport/nvls.cc. :1618 uses bare NCCLCHECK, not NCCLCHECKGOTO, so a failure in ncclNvlsInit
+// returns WITHOUT running exit: -- the counter on ncclOsCpuCount is what makes that bypass observable.
+extern ncclResult_t g_ncclNvlsInitResult;
+extern int g_ncclNvlsInitCalls;
+// init.cc:2185, gated on comm->nvlsSupport surviving the :2182 fold.
+extern ncclResult_t g_ncclNvlsTuningResult;
+extern int g_ncclNvlsTuningCalls;
 
 void ResetTransportStubs();
